Add setup related files
Add all setup and project related files. Chose to include both the
pyproject.toml, and a setup.py, as `pip install -e .` does not work
without a setup.py.

Added an initial change log. The current version of the code works for
simple use cases, and can be released as 0.1.0.